+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migration of mongoClient, buildUrlMap and mqttPublisher to TS
 - Migration of authMiddleware and socketConfig to TS
 - Migration of rateLimiter and getTaskEnabler to TS
+- Migration of switchTaskEnabler to TS
+
+### Added
+- id and name field for frontend
 
 ## [v15.3.0] 2023-12-08
 ### Removed